Move actions to client.internal.request.actions
