+++ Client/src/main/java/de/turtle_exception/client/internal/request/actions/NetAction.java
@@ -1,9 +1,10 @@
-package de.turtle_exception.client.internal.request;
+package de.turtle_exception.client.internal.request.actions;
 
 import de.turtle_exception.client.internal.ActionImpl;
 import de.turtle_exception.client.internal.Provider;
 import de.turtle_exception.client.internal.data.Data;
 import de.turtle_exception.client.internal.net.Connection;
+import de.turtle_exception.client.internal.request.*;
 import kotlin.NotImplementedError;
 import org.jetbrains.annotations.NotNull;
