+++ Client/src/main/java/de/turtle_exception/client/internal/request/actions/ProxyAction.java
@@ -1,4 +1,4 @@
-package de.turtle_exception.client.internal.request;
+package de.turtle_exception.client.internal.request.actions;
 
 import de.turtle_exception.client.api.request.Action;
 import de.turtle_exception.client.internal.ActionImpl;
